pyproject.toml
[project]
name = "smart-genai-bot"
version = "0.1.0"
description = "Citation-aware Agentic RAG chatbot built with LangGraph, LlamaIndex, OpenAI Responses API, FastAPI, and Uvicorn."
readme = "README.md"
requires-python = ">=3.12,<3.14"
license = { text = "MIT" }
authors = [
  { name = "Mei, Tathagata, Zaman" }
]
keywords = [
  "rag",
  "agentic-rag",
  "langgraph",
  "llamaindex",
  "openai",
  "fastapi",
  "chatbot",
  "citations"
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]

dependencies = [
  "fastapi>=0.115.0",
  "langchain>=0.3.0",
  "langgraph>=0.2.0",
  "llama-index>=0.12.0",
  "llama-index-readers-file>=0.4.0",
  "openai>=1.50.0",
  "pydantic>=2.9.0",
  "pypdf>=5.0.0",
  "python-dotenv>=1.0.1",
  "uvicorn>=0.30.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.6.0"
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"